requirements.txt
transformers
torch
faiss-cpu
redis
twilio
gtts
reportlab
requests
python-dotenv
sentence-transformers
